Implement FI_FENCE
Implement FI_FENCE flag for RMA operations.  Outstanding TX count of each VC is
tracked.  When a request with FI_FENCE specified arrives at the front of the TX
queue, it is not processed until the TX count drops to zero.  All subsequent
requests are also queued until that time.  This code should apply to all TX
request types, not just RMA.

Signed-off-by: Zach Tiffany <ztiffany@cray.com>