prov/gni: provide way for app to set cdm_id
This mod provides a way for an application to
control the cdm_id (and hence the Aries address)
of FI_EP_RDM endpoints it opens.  Specifically,
by providing a non-NULL node and service argument
to fi_getinfo, and setting the flag to FI_SOURCE,
the GNI provider will assign a CDM_ID equal to the
numerical value of the service argument.

Modified one of the datagram criterion tests to
validate this feature.

Signed-off-by: Howard Pritchard <howardp@lanl.gov>