prov/gni: various shutdown related fixes
Turns out some of the criterion tests which intentionally
do things to check for errors, etc. have a tendency to
bring out some issues in the destruction of objects within
the GNI provider when EPs, domains, etc. are closed and
there are outstanding transactions/GNI RX CQEs etc. still
outstanding.

One of the ways that these shutdowns with GNI objects still
queued up (like CQEs on GNI hw cqs) was seg faults when
a thread tried to process CQEs that were associated with a VC
that was being destroyed by another thread.  The most frequent
signature was a thread trying to access the flag field of a VC
struct in, for example _gnix_vc_rx_schedule.

Currently it is criterion tests which are trying to check edge
cases that hit these kind of problems.

use a nic based freelist for vc's

Signed-off-by: Howard Pritchard <howardp@lanl.gov>