prov/gnix: Reference counting with MR system
NICs allocated from the context of a memory registration start with
a single reference that is never dropped because a second reference
is taken by the original registration. gnix_nic_alloc assumes the
calling function will take the reference provided by the function
and that assumption wasn't being observed.

This mod fixes the issue by only taking a reference if a NIC wasn't
created during the memory registration request.

Signed-off-by: James Swaro <jswaro@cray.com>