prov/gni: Fix problems with GNI cm nic progress
It turns out there are certain kGNI datagram exchange
patterns that result in the current GNI provider
CM NIC mechanism not making forward progress when
FI_PROGRESS_AUTO is requested for control progress.

This situation can arise when multiple datagrams
using the same Aries NIC/cookie/cdm_id are being processed
by the progress thread.  The underlying problem is that the
kGNI datagram mechanism only allows one outstanding datagram
for a given target nic/cookie/cdm_id to be posted to the
state engine at a time, and the fact that there is no notification
to the application when a previously posted, bound datagram
transaction completes.  As a consequence, there can be cases where
a process posts a series of connection requests to a target process
with same cookie/cdm_id, that results in the target responding
to the first, then getting GNI_RC_RESOURCE_ERROR for the second because
the first has not completed.

The first response completes, but the progress thread is not woken
up, i.e. does not break out of GNI_PostdataProbeWaitById to progress
the backlog of connection management responses.

We do not see this problem with MPI tests nor simple OpenSHMEM tests
because each process is only using a single endpoint, and thus its
not possible to get this multiple connection requests with same
target aries/cookie/cdm_id.

However we can see this with both shared TXs and with scalable EPs
since in both cases we can have multiple EP's using the same gnix_cm_nic.
Shared TX's are being used in a new version of SOS OpenSHMEM, and
that's where this problem was observed.

One way to fix this problem would be to have a timeout on the
GNI_PostdataProbeWaitById and add a method for the progress thread to progress
backlogged connection management requests.  This has the disadvantage that
we are adding noise by having the progress thread periodically wake up.

In this PR, an alternate solution is implemented where, as part of the general
gnix_cm_nic progress mechanism, we check the connection management request
list regardless of whether the application has requested FI_PROGRESS_AUTO or
not for control progress.

In addressing this issue,  another problem was found with the progression
mechanism.  Namely, we do need to progress the GNI provider state when an
application executes a fi_cq_readerr on a GNI provider CQ.  See the sockets
provider for the correct way to do this.

Signed-off-by: Howard Pritchard <howardp@lanl.gov>