prov/gni: be more accurate about nic allocation
when using shared TX contexts.

Also, improve the cleanup code in the error path
when we fail to allocate a gnix_nic due to underlying
resource constraints.

Also, since they chew up IOMMU space and aren't being
used currently, if def out code to allocate send/receive
buffers when allocating a gnix_nic.

Adjust allocator test to take account of the fact
that we're no longer allocating slabs when a nic is
allocated.

Signed-off-by: Howard Pritchard <howardp@lanl.gov>