prov/gni: fix gnix wait stalls
The wait subsystem could deadlock with the MR
code as they both took the gnix_nic_list_lock.
Another cause of deadlock was gni can issue
completions so fast that the fd becomes full.

Refactored wait_thread_func to build a list
with a ref on the nic then iterate over those
nic's without the lock.

Change wait write_fd to be non_blocking so
when the pipeline is full it doesnt get stuck.

Signed-off-by: James Shimek <jshimek@cray.com>