prov/gni: Restrict mr mode bits to basic
For the GNI provider, only basic memory registration is supported at
this time. This commit fixes the provider to only use basic memory
registration.

In addition, this commit resolves problems with the GNI provider
on fi_getinfo where the GNI fi_getinfo equivalent function call
would modify the hints structure, which is not permitted.

Requires patch for ofiwg/libfabric#3037

Signed-off-by: James Swaro <jswaro@cray.com>