prov/gni: add version check FI_NOTIFY_FLAGS_ONLY
FI_NOTIFY_FLAGS_ONLY is only allowed when used with
applications requesting version 1.5 or greater of the
gni provider. Remove that mode when returning the info
struct when returning from fi_getinfo.

Add tests that verify the fi_info->mode returned by the
provider masks off FI_NOTIFY_FLAGS_ONLY when
version requested is less than 1.5 and is not masked
off for 1.5 and greater.

Signed-off-by: James Shimek <jshimek@cray.com>