prov/gni: fix a bind issue with SEPs
An SEP can be bound to an AV prior to or after
the SEP is used to create tx/rx contexts.

This commit allows for the first case to work.

Also remove erroneous code allowing bind of counter
objects to scalable eps.  This is not supported according
to the fi_scalable_ep_bind description in the fi_endpoint
man page.

Fixes #ofi-cray/libfabric-cray#1386

Signed-off-by: Howard Pritchard <howardp@lanl.gov>