prov/gni: add ability to detect local PE rank
There may be some cases, when the GNI provider is being
used in the context of a job launcher like SLURM or ALPS,
that the provider may need to use the local rank of
a process in deciding how to allocate Aries HW resources
to the process.

This commit adds a function to determine the "local rank".
This doesn't apply to processes not started via a SLURM
or ALPS based launch.

Signed-off-by: Howard Pritchard <howardp@lanl.gov>