prov/gni: fix issue with FI_SOURCE_ERR
Fix problems with use of FI_SOURCE/FI_SOURCE_ERR.
Fixes ofi-cray/libfabric-cray#1409

Part of the reason this problem wasn't detected was that
Criterion was not covering it for several reasons including
1) tests for FI_SOURCE/FI_SOURCE_ERR were not enabled
2) changes to better test FI_MULTIRECV conflicted with testing
FI_SOURCE/FI_SOURCE_ERR

To deal with this, activate the tests in rdm_sr.c and moved
the multireceive specific tests to a separate file just for
multirecv testing.

A future PR may deal with cleanup of redundant code in the Criterion
tests.

Signed-off-by: Howard Pritchard <howardp@lanl.gov>