prov/gni: fix a problem with internal rcache
The GNI provider internal rcache had a bug when
trying to compute overlaps of incoming memory
registration requests with existing registrations
in the rcache.

The nature of the problem was such that it was almost
never encountered with OpenSHMEM or MPI applications,
but for Mercury, which was doing consecutive registrations
of buffers within a larger previously allocated virtual
memory range, the bug was hit.

Also add some better debug output for the MR subsys in
the GNI provider.

Thanks to Jerome Soumagne for reporting.

@soumagne

Signed-off-by: Howard Pritchard <howardp@lanl.gov>